Fix: compatibilidad redux
